backend/requirements.txt
fastapi
uvicorn
pandas
openpyxl
python-multipart
sqlalchemy
psycopg2-binary
python-dotenv
